[FINAL PROJECT] Enhance chat features with media support, message handling, and UI updates
Some dude in comments said it's final... 😭

+++ src/Whisper.Client/src/api/agent.ts
@@ -4,6 +4,8 @@ import { getAuthTokenFromDB, clearAuthData } from "./db";
 import type { UpdateUserDto, UserDto } from "../types/user";
 import type { EmojiDto } from "../types/emoji";
 import type { BinlistResponseDto } from "../types/binlist";
+import type { GifResponseDto } from "../types/gif";
+import type { ChatDto } from "../types/chat";
 
 const agent = axios.create({
     baseURL: import.meta.env.VITE_API_URL,
@@ -30,7 +32,7 @@ agent.interceptors.request.use(async (config: InternalAxiosRequestConfig) => {
 agent.interceptors.response.use(
     (response) => response,
     async (error) => {
-        if (error.response && error.response.status === 401) {
+        if (error.response && error.response.status === 401 && window.location.pathname !== "/auth/login") {
             await clearAuthData();
             localStorage.clear();
             window.location.replace("/auth/login");
@@ -59,7 +61,7 @@ const Auth = {
 };
 
 const Chats = {
-    list: () => requests.get<any[]>("/api/v1/Chats/list"), 
+    list: () => requests.get<ChatDto[]>("/api/v1/Chats/list"), 
     messages: (chatId: string) => 
         requests.get<any[]>(`/api/v1/Chats/${chatId}/messages?limit=50&offset=0`),
     create: (receiverId: string, name: string) => 
@@ -69,7 +71,8 @@ const Chats = {
         }),
     sendMessage: (body: any) => 
         requests.post<any>("/api/v1/Chats/messages", body),
-    getMembers: (chatId: string) => requests.get<string[]>(`/api/v1/Chats/${chatId}/get-members`),
+    getMembers: (chatId: string) => requests.get<UserDto[]>(`/api/v1/Chats/${chatId}/get-members`),
+    remove: (chatId: string) => requests.delete(`/api/v1/Chats/${chatId}`)
 };
 
 const Users = {
@@ -98,11 +101,19 @@ const Keys = {
 
 const Emojis = {
     getAll: () =>
-        requests.get<any>("/api/v1/Emojis/"),
+        requests.get<EmojiDto[]>("/api/v1/Emojis/"),
 
     getBySearch: (query: string) => 
-        requests.get<EmojiDto[]>(`/api/v1/Emojis/search?query=${query}`)
+        requests.get<EmojiDto[]>(`/api/v1/Emojis/search?query=${query}`),
 };
+
+const Gifs = {
+    getTrending: () =>
+        requests.get<GifResponseDto>("/api/v1/Gifs/trending"),
+
+    getBySearch: (query: string) =>
+        requests.get<GifResponseDto>(`/api/v1/Gifs/search?query=${query}`),
+}
 
 const Enrichment = {
     getCardData: (bin: string) => requests.get<BinlistResponseDto>(`/api/v1/Binlist/card/${bin}`),
@@ -111,6 +122,19 @@ const Admin = {
     getDashboardData: () => requests.get<any>("/api/v1/Admin/dashboard"),
 };
 
-const agentService = { Chats, Users, Auth, Keys, Enrichment, Admin, Emojis };
+const Media = {
+    upload: (file: File) => {
+        const formData = new FormData();
+        formData.append("file", file); 
+        return agent.post<any>("/api/v1/Media/upload", formData);
+    },
+    download: (fileId: string) => {
+        return agent.get<Blob>(`/api/v1/Media/download/${fileId}`, {
+            responseType: 'blob'
+        }).then(response => response.data);
+    }
+}
+
+const agentService = { Chats, Users, Auth, Keys, Enrichment, Admin, Emojis, Gifs, Media };
 
 export default agentService;
